Move path manipulation tools to a dedicated object.
And clean the imports in test sources.
